requirements/common.txt
Django==1.11.4

Jinja2==2.9.6
django-jinja==2.3.1
django-jinja2==0.1

Pillow==4.2.1

whitenoise==3.3.0

pytz==2016.10
python-dateutil==2.6.0

django-environ==0.4.3

psycopg2